# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Aniwatch API
  version: 2.18.2
  description: |
    Unofficial REST API for anime data scraped from hianimez.to via the aniwatch library.
servers:
  - url: http://localhost:3000
    description: Local dev server

components:
  parameters:
    AniwatchCacheExpiry:
      name: Aniwatch-Cache-Expiry
      in: header
      required: false
      description: Optional cache TTL in seconds (only used if Redis caching is enabled).
      schema:
        type: integer
        minimum: 1
  responses:
    GenericSuccess:
      description: Success
      headers:
        Aniwatch-Cache-Expiry:
          description: Cache TTL in seconds, echoed only when Redis caching is enabled.
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericSuccess'
    ErrorResponse:
      description: Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    GenericSuccess:
      type: object
      properties:
        status:
          type: integer
          example: 200
        data:
          type: object
          additionalProperties: true
      required:
        - status
        - data
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          example: 500
        message:
          type: string
      required:
        - status
        - message

paths:
  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: string
                example: daijoubu

  /app-version:
    get:
      summary: App version
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: string
                example: 1.0.4

  /update:
    get:
      summary: Update link
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: string
                example: https://drive.google.com/file/d/1DrES5qI7wVtj3TKrN-3tMRAqR2K3enKV/view?usp=drive_link

  /v:
    get:
      summary: Version info
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: string

  /api/v2/anicrush:
    get:
      summary: Placeholder for future Anicrush endpoint
      parameters:
        - $ref: '#/components/parameters/AniwatchCacheExpiry'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: string

  /api/v2/hianime/:
    get:
      summary: Redirect to root
      parameters:
        - $ref: '#/components/parameters/AniwatchCacheExpiry'
      responses:
        '301':
          description: Moved Permanently

  /api/v2/hianime/home:
    get:
      summary: Get home page data
      parameters:
        - $ref: '#/components/parameters/AniwatchCacheExpiry'
      responses:
        '200':
          $ref: '#/components/responses/GenericSuccess'
        '500':
          $ref: '#/components/responses/ErrorResponse'

  /api/v2/hianime/azlist/{sortOption}:
    get:
      summary: Get A-Z list
      parameters:
        - name: sortOption
          in: path
          required: true
          schema:
            type: string
          description: Sort option for A-Z list.
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - $ref: '#/components/parameters/AniwatchCacheExpiry'
      responses:
        '200':
          $ref: '#/components/responses/GenericSuccess'
        '500':
          $ref: '#/components/responses/ErrorResponse'

  /api/v2/hianime/qtip/{animeId}:
    get:
      summary: Get Qtip info
      parameters:
        - name: animeId
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/AniwatchCacheExpiry'
      responses:
        '200':
          $ref: '#/components/responses/GenericSuccess'
        '500':
          $ref: '#/components/responses/ErrorResponse'

  /api/v2/hianime/category/{name}:
    get:
      summary: Get anime by category
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - $ref: '#/components/parameters/AniwatchCacheExpiry'
      responses:
        '200':
          $ref: '#/components/responses/GenericSuccess'
        '500':
          $ref: '#/components/responses/ErrorResponse'

  /api/v2/hianime/genre/{name}:
    get:
      summary: Get anime by genre
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - $ref: '#/components/parameters/AniwatchCacheExpiry'
      responses:
        '200':
          $ref: '#/components/responses/GenericSuccess'
        '500':
          $ref: '#/components/responses/ErrorResponse'

  /api/v2/hianime/producer/{name}:
    get:
      summary: Get anime by producer
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - $ref: '#/components/parameters/AniwatchCacheExpiry'
      responses:
        '200':
          $ref: '#/components/responses/GenericSuccess'
        '500':
          $ref: '#/components/responses/ErrorResponse'

  /api/v2/hianime/schedule:
    get:
      summary: Get estimated schedule
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: string
          description: Date string passed to the scraper.
        - name: tzOffset
          in: query
          required: false
          schema:
            type: integer
            default: -330
          description: Timezone offset in minutes.
        - $ref: '#/components/parameters/AniwatchCacheExpiry'
      responses:
        '200':
          $ref: '#/components/responses/GenericSuccess'
        '500':
          $ref: '#/components/responses/ErrorResponse'

  /api/v2/hianime/search:
    get:
      summary: Search anime
      description: Additional filter query params are passed through to the scraper.
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
          description: Search query.
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - $ref: '#/components/parameters/AniwatchCacheExpiry'
      responses:
        '200':
          $ref: '#/components/responses/GenericSuccess'
        '500':
          $ref: '#/components/responses/ErrorResponse'

  /api/v2/hianime/search/suggestion:
    get:
      summary: Get search suggestions
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
        - $ref: '#/components/parameters/AniwatchCacheExpiry'
      responses:
        '200':
          $ref: '#/components/responses/GenericSuccess'
        '500':
          $ref: '#/components/responses/ErrorResponse'

  /api/v2/hianime/anime/{animeId}:
    get:
      summary: Get anime info
      parameters:
        - name: animeId
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/AniwatchCacheExpiry'
      responses:
        '200':
          $ref: '#/components/responses/GenericSuccess'
        '500':
          $ref: '#/components/responses/ErrorResponse'

  /api/v2/hianime/anime/{animeId}/episodes:
    get:
      summary: Get anime episodes
      parameters:
        - name: animeId
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/AniwatchCacheExpiry'
      responses:
        '200':
          $ref: '#/components/responses/GenericSuccess'
        '500':
          $ref: '#/components/responses/ErrorResponse'

  /api/v2/hianime/anime/{animeId}/next-episode-schedule:
    get:
      summary: Get next episode schedule
      parameters:
        - name: animeId
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/AniwatchCacheExpiry'
      responses:
        '200':
          $ref: '#/components/responses/GenericSuccess'
        '500':
          $ref: '#/components/responses/ErrorResponse'

  /api/v2/hianime/episode/servers:
    get:
      summary: Get episode servers
      parameters:
        - name: animeEpisodeId
          in: query
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/AniwatchCacheExpiry'
      responses:
        '200':
          $ref: '#/components/responses/GenericSuccess'
        '500':
          $ref: '#/components/responses/ErrorResponse'

  /api/v2/hianime/episode/sources:
    get:
      summary: Get episode sources
      parameters:
        - name: animeEpisodeId
          in: query
          required: true
          schema:
            type: string
        - name: server
          in: query
          required: false
          schema:
            type: string
          description: Server name, defaults to VidStreaming.
        - name: category
          in: query
          required: false
          schema:
            type: string
            enum:
              - sub
              - dub
              - raw
          description: Content category.
        - $ref: '#/components/parameters/AniwatchCacheExpiry'
      responses:
        '200':
          $ref: '#/components/responses/GenericSuccess'
        '500':
          $ref: '#/components/responses/ErrorResponse'

  /api/v2/manga:
    get:
      summary: List supported manga providers
      parameters:
        - $ref: '#/components/parameters/AniwatchCacheExpiry'
      responses:
        '200':
          $ref: '#/components/responses/GenericSuccess'

  /api/v2/manga/mangahere/{query}:
    get:
      summary: Search manga on MangaHere
      parameters:
        - name: query
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
        - $ref: '#/components/parameters/AniwatchCacheExpiry'
      responses:
        '200':
          $ref: '#/components/responses/GenericSuccess'

  /api/v2/manga/mangahere/rankings:
    get:
      summary: Get MangaHere rankings
      parameters:
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum:
              - total
              - month
              - week
              - day
            default: total
        - $ref: '#/components/parameters/AniwatchCacheExpiry'
      responses:
        '200':
          $ref: '#/components/responses/GenericSuccess'
        '400':
          $ref: '#/components/responses/ErrorResponse'

  /api/v2/manga/mangahere/hot:
    get:
      summary: Get MangaHere hot releases
      parameters:
        - $ref: '#/components/parameters/AniwatchCacheExpiry'
      responses:
        '200':
          $ref: '#/components/responses/GenericSuccess'

  /api/v2/manga/mangahere/trending:
    get:
      summary: Get MangaHere trending manga
      parameters:
        - $ref: '#/components/parameters/AniwatchCacheExpiry'
      responses:
        '200':
          $ref: '#/components/responses/GenericSuccess'

  /api/v2/manga/mangahere/recent-updates:
    get:
      summary: Get MangaHere recent updates
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - $ref: '#/components/parameters/AniwatchCacheExpiry'
      responses:
        '200':
          $ref: '#/components/responses/GenericSuccess'

  /api/v2/manga/mangahere/home:
    get:
      summary: Get MangaHere home data
      description: Returns `top10Today` and `top10Weekly` (limited to 10 each), plus full `hot`, `trending`, and `recentUpdates` arrays from first page.
      parameters:
        - $ref: '#/components/parameters/AniwatchCacheExpiry'
      responses:
        '200':
          $ref: '#/components/responses/GenericSuccess'

  /api/v2/manga/mangahere/info:
    get:
      summary: Get MangaHere manga info
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/AniwatchCacheExpiry'
      responses:
        '200':
          $ref: '#/components/responses/GenericSuccess'
        '400':
          $ref: '#/components/responses/ErrorResponse'

  /api/v2/manga/mangahere/read:
    get:
      summary: Get MangaHere chapter pages (manga sources)
      parameters:
        - name: chapterId
          in: query
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/AniwatchCacheExpiry'
      responses:
        '200':
          $ref: '#/components/responses/GenericSuccess'
        '400':
          $ref: '#/components/responses/ErrorResponse'